Make sure active model compatible classes return true/false for create/save
Calling `create` or `save` on a resource returns the object. This
causes problems for ActiveModel compatible resources using
`acts_as_active_model`, because Rails expects `create` and `save` to
return a boolean value.
This commit changes the behaviour, `save` and `create` now return true
if the action was successful or false if there were any validation
errors.