Only get needed attributes for sanitizing the url
Previously, when sanitizing the url, we called `to_h` on the object to
get all the available attributes to substitute in the url. This could
cause api requests in the background if there were any overwritten
attribute getters (e.g. to map nested resources as objects). This
commit changes the sanitizer to only get needed attributes for
sanitizing.